more styling + sidebar in prep room
