+++ app/public/src/pages/lobby.css
@@ -6,6 +6,7 @@
 
 .lobby-container {
   padding: 10px;
+  width: 100%;
 }
 
 .lobby nav {
@@ -45,13 +46,24 @@
 }
 
 .lobby .room-menu {
-  width: 480px;
   height: 100%;
+  min-width: 25%;
+  font-size: 16px;
+}
+
+.main-lobby ul {
+  display: flex;
+  font-size: 16px;
+  overflow-x: scroll;
 }
 
 .lobby .current-users-menu {
   min-width: 333px;
   height: 100%;
+}
+
+.lobby .current-users-menu ul {
+  display: unset;
 }
 
 .lobby .user-chat {
@@ -104,10 +116,6 @@
   height: 100%;
 }
 
-.main-lobby > div {
-  min-width: 405px;
-}
-
 .main-lobby .news-chat {
   display: flex;
   gap: 4px;
@@ -119,37 +127,7 @@
   }
 }
 
-.menu-item {
-  width: 100%;
-}
-
-.menu-item a {
-  width: 100%;
-  cursor: var(--cursor-hover) !important;
-}
-
-.menu-item a span {
-  display: flex;
-  gap: 8px;
-  align-items: center;
-}
-
-.ps-menu-root ul {
-  width: 100%;
-}
-
 .lobby-container {
   padding: 20px;
 }
 
-.icon {
-  position: relative;
-}
-
-.notification {
-  position: absolute;
-  top: 0;
-  right: -12px;
-  height: 6px;
-  font-size: 6px;
-}
