+++ app/public/src/pages/lobby.tsx
@@ -1,16 +1,13 @@
 import React, { useCallback, useEffect, useState } from "react"
 import { Navigate } from "react-router-dom"
 import Chat from "./component/chat/chat"
-import News from "./component/news/news"
 import CurrentUsers from "./component/available-user-menu/current-users"
 import RoomMenu from "./component/available-room-menu/room-menu"
 import TabMenu from "./component/lobby-menu/tab-menu"
 import firebase from "firebase/compat/app"
 import { FIREBASE_CONFIG } from "./utils/utils"
 import { Room, RoomAvailable } from "colyseus.js"
-import Wiki from "./component/wiki/wiki"
 import TeamBuilder from "./component/bot-builder/team-builder"
-import MetaReport from "./component/meta-report/meta-report"
 import { useAppDispatch, useAppSelector } from "../hooks"
 import {
   joinLobby,
@@ -58,15 +55,12 @@ import LobbyUser from "../../../models/colyseus-models/lobby-user"
 import { IBot } from "../../../models/mongo-models/bot-v2"
 import { IMeta } from "../../../models/mongo-models/meta"
 import { IItemsStatistic } from "../../../models/mongo-models/items-statistic"
-import PokemonCollection from "./component/collection/pokemon-collection"
 import PokemonConfig from "../../../models/colyseus-models/pokemon-config"
-import Booster from "./component/booster/booster"
 import { IPokemonsStatistic } from "../../../models/mongo-models/pokemons-statistic"
 import "./lobby.css"
 import { BotManagerPanel } from "./component/bot-builder/bot-manager-panel"
 import i18n from "../i18n"
-import { Header } from "./header"
-import { BasicModal } from "./component/modal/modal"
+import { MainSidebar } from "./main-sidebar"
 
 export type Page = "main_lobby" | "bot_builder" | "bot_manager"
 
@@ -81,7 +75,6 @@ export default function Lobby() {
 
   const [lobbyJoined, setLobbyJoined] = useState<boolean>(false)
   const [page, setPage] = useState<Page>("main_lobby")
-  const [modal, setModal] = useState<Modals>()
 
   const [toPreparation, setToPreparation] = useState<boolean>(false)
   const [toAuth, setToAuth] = useState<boolean>(false)
@@ -94,10 +87,6 @@ export default function Lobby() {
   }, [lobbyJoined, dispatch, client])
 
   const changePage = useCallback((nextPage: Page) => setPage(nextPage), [])
-  const changeModal = useCallback(
-    (nextModal: Modals) => setModal(nextModal),
-    []
-  )
 
   if (toAuth) {
     return <Navigate to={"/"} />
@@ -109,14 +98,11 @@ export default function Lobby() {
 
   return (
     <main className="lobby">
-      <Header
+      <MainSidebar
         changePage={changePage}
-        changeModal={changeModal}
         showBackButton={page !== "main_lobby"}
       />
       <div className="lobby-container">
-        <Modals modal={modal} setModal={setModal} />
-
         {page === "bot_manager" && user?.role !== Role.BASIC && (
           <BotManagerPanel />
         )}
@@ -148,64 +134,7 @@ function MainLobby({ toPreparation, setToPreparation }) {
   )
 }
 
-export type Modals = "meta" | "wiki" | "collection" | "booster" | "news"
 
-function Modals({
-  modal,
-  setModal
-}: {
-  modal?: Modals
-  setModal: (nextModal?: Modals) => void
-}) {
-  const {
-    meta,
-    metaItems,
-    metaPokemons
-  }: Partial<INetwork> & Partial<IUserLobbyState> = useAppSelector((state) => ({
-    meta: state.lobby.meta,
-    metaItems: state.lobby.metaItems,
-    metaPokemons: state.lobby.metaPokemons
-  }))
-
-  const closeModal = useCallback(() => setModal(undefined), [setModal])
-
-  return (
-    <>
-      <BasicModal
-        handleClose={closeModal}
-        show={modal === "news"}
-        body={<News />}
-      />
-      <BasicModal
-        handleClose={closeModal}
-        show={modal === "collection"}
-        body={<PokemonCollection />}
-      />
-      <BasicModal
-        handleClose={closeModal}
-        show={modal === "booster"}
-        body={<Booster />}
-      />
-      <BasicModal
-        handleClose={closeModal}
-        show={modal === "wiki"}
-        body={<Wiki />}
-      />
-      <BasicModal
-        title={"meta"}
-        show={modal === "meta" && meta.length > 0 && metaItems.length > 0}
-        handleClose={closeModal}
-        body={
-          <MetaReport
-            meta={meta}
-            metaItems={metaItems}
-            metaPokemons={metaPokemons}
-          />
-        }
-      />
-    </>
-  )
-}
 
 const join = async (dispatch, client, setToAuth) => {
   if (!firebase.apps.length) {
